+++ pyproject.toml
@@ -1,0 +1,2 @@
+[tool.ruff]
+line-length = 120
