+++ tests/gui/test_plot_view.py
