+++ utils/__init__.py
@@ -1,3 +1,2 @@
 from .data_utils import one_hot_encode, validate_csv
 from .logging_config import setup_logging
-from .plotting_utils import PlottingWidget
